requirements.txt
streamlit>=1.30.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
pypdf>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
tavily-python>=0.5.0
